Get lines drawing
I fixed a bug in the cache where it was calling `downcast_ref` on the
Any in an Arc, which failed because that Any was a GliumStrokeCache, not
an Arc<GliumStrokeCache>. Now there's an extra box so that the
Arc<GliumStrokeCache> can be coerced to Any instead of just the
GliumStrokeCache. Might be able to replace the Box with another type.

Now line path segments are converted to triangles and the Path::move_to
and Path::line_to methods add segments. Also, now stencil_stroke_path
takes a matrix to transform the path by.