Get quadratic Bezier curves rendering in a fragment shader
Unfortunately, antialiasing isn't supported yet. It may be possible to
set gl_SampleMask in the fragment shader (which would be kind of slow
and I'd need to pass the sample positions to the shader somehow).
Alternatively, I could just run the fragment shader for each sample,
which might even be faster.