Work on sketching out the line layout algorithm more
I've added a paragraph direction, which is the overall layout direction
of text. And handling runs that are a different direction than the
paragraph direction is more complicated than I realized. Now I have
measure_line() tracking the total width of runs with the same direction
because it's impossible to layout mixed direction runs otherwise
(because a later run that is a different direction can affect the
position of an earlier run).

I've also added a test that fails and added a long list of tests that
need to be added.